+++ src/Command/Install/CreateVendorsLinksCommand.php
@@ -18,8 +18,8 @@ namespace MeTools\Command\Install;
 use Cake\Console\Arguments;
 use Cake\Console\ConsoleIo;
 use Cake\Console\ConsoleOptionParser;
-use Cake\Core\Configure;
 use MeTools\Command\Command;
+use MeTools\Core\Configure;
 use Tools\Filesystem;
 
 /**
@@ -46,11 +46,13 @@ class CreateVendorsLinksCommand extends Command
      */
     public function execute(Arguments $args, ConsoleIo $io): void
     {
+        $links = array_merge(...array_values(Configure::readFromPlugins('VendorLinks')));
+
         $Filesystem = new Filesystem();
-        foreach (array_unique(Configure::readOrFail('VENDOR_LINKS')) as $origin => $target) {
+        foreach ($links as $origin => $target) {
             $this->createLink(
                 $io,
-                $Filesystem->concatenate(ROOT, 'vendor', $origin),
+                $Filesystem->concatenate(ROOT, 'vendor', $Filesystem->normalizePath($origin)),
                 $Filesystem->concatenate(WWW_ROOT, 'vendor', $target)
             );
         }