+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 ## 2.23 branch
 ### 2.23.3
 * added `Configure` class, with `readFromPlugins()` method;
-* uses `Configure::readFromPlugins()` with `WritableDirs` key instead of the old `WRITABLE_DIRS` key.
+* uses `Configure::readFromPlugins()` with `WritableDirs` and `VendorLinks` keys instead of the old `WRITABLE_DIRS` and 
+  `VENDOR_LINKS` keys.
 
 ### 2.23.2
 * fixed a little bug for `RunAllCommand` with the `--force` option.